+++ src/core/error.service.ts
@@ -1,4 +1,4 @@
-/**
+/***
  * Copyright (c) 2017 [Arthur Xie]
  * <https://github.com/kite-js/kite>
  * 
@@ -20,8 +20,8 @@ export class ErrorService {
 
     /**
      * 
-     * @param code 错误代码
-     * @param extra 附加错误消息
+     * @param code error code
+     * @param extra error extra message
      */
     getError(code: number | string, extra ?: string | string[]) {